Crate sparse_merkle

This module implements a Sparse Merkle Trie backed by storage module. The trie itself doesn't persist anything, but realizes the logic of trie R/W only. The write path will produces intermediate results in cache for storage layer to commit and the read path will return results directly. The public APIs are only new, put_write_set and get_with_proof. After each put with a write_set based on a known root, the trie module will return a new root with new generated trie nodes and blobs in a batch for storage module to commit.

The sparse merkle trie itself logically is a 256-bit merkle trie with an optimization that any subtree containing 0 or 1 leaf node will be replaced by that leaf node or a dummy node with default hash value. With this optimization we could save CPU a lot by avoid hashing on many sparse levels in the tree. Physically, the trie is implemented as a modified pactricia merkle trie like ethereum with some modifications in hashing scheme to match the optimization just stated.

Modules

node_type
Node types of modified sparse merkle tree

Structs

SparseMerkleTrie

TrieUpdateBatch
TrieUpdateBatch is a wrapper of NodeBatch and BlobBatch that represents the incremental updates of trie after applying a write set, which is a vector of account_address and new_account_blob pairs.

Traits

TrieReader
TrieReader defines the interface between SparseMerkleTrie and underlying storage holding nodes and blobs.

Type Definitions

BlobBatch
Blob batch that will be written into db atomically with other batches

NodeBatch
Node batch that will be written into db atomically with other batches