Struct sparse_merkle::node_type::BranchNode

pub struct BranchNode { /* fields omitted */ }

A branch node represents a 4-level subtree with 16 children 4 levels down. The introduction of branch node could benefit IOPS of querying a tree by 4x since we compress 4 levels in standard merkle tree into 1 node. Though we choose the same branch node structure as that of patricia merkle tree, the merkle proof has the same format as that of modified sparse merkle tree, the root hash computation logic is similar to a 4-level sparse merkle tree except some customizations. See the CryptoHash trait implementation below for details.

Methods

impl BranchNode

pub fn new(children: HashMap<u8, (HashValue, bool)>) -> Self

Create a new branch node

pub fn set_child(&mut self, n: u8, child: (HashValue, bool))

Set the n-th child to given hash and store whether it is leaf.

pub fn child(&self, n: u8) -> Option<HashValue>

Get the hash of the n-th child.

pub fn is_leaf(&self, n: u8) -> Option<bool>

Whether the n-th child is leaf node or not.

pub fn num_children(&self) -> usize

Return the total number of children

pub fn generate_bitmaps(&self) -> (u16, u16)

Generate child_bitmap and leaf_bitmap: child at index i exists if child_bitmap[i] is set; child at index i is leaf node if leaf_bitmap[i] is set.

pub fn merkle_hash(
&self,
start: u8,
width: u8,
(child_bitmap, leaf_bitmap): (u16, u16)
) -> HashValue

Compute the hash of branch node according to modifed sparse merkle tree data structure in the logical view. start and nibble_height define a subtree that we want a root hash from. For a branch node with 16 children at height 0, we compute the hash of branch node as if a full binary merkle tree with 16 leaves as below:

 
  4 ->              +------ root hash ------+
                    |                       |
  3 ->        +---- # ----+           +---- # ----+
              |           |           |           |
  2 ->        #           #           #           #
            /   \       /   \       /   \       /   \
  1 ->     #     #     #     #     #     #     #     #
          / \   / \   / \   / \   / \   / \   / \   / \
  0 ->   0   1 2   3 4   5 6   7 8   9 A   B C   D E   F
  ^
height

As illustrated above, at nibble height 0, 0..F in hex denote 16 chidren hashes. Each '#' means the hash of its two direct children, which will be used to generate the hash of its parent. Finally, we can get the digest of this branch node.

However, if branch node doesn't have 16 chidren at height 0 but just few of them, we have a modified hashing rule on top of what stated above:

From top to bottom, a node will be replaced by a leaf child if the subtree rooted at this node has only one child at height 0 and it is the leaf child.

From top to bottom, a node will be replaced by the default node if the subtree rooted at this node doesn't have any child at height 0. For example, if a branch node has 3 leaf nodes at index 0, 3, 8, respectively, and 1 branch/extension node at index C, then the computation graph will be like:

 
  4 ->              +------ root hash ------+
                    |                       |
  3 ->        +---- # ----+           +---- # ----+
              |           |           |           |
  2 ->        #           @           8           #
            /   \                               /   \
  1 ->     0     3                             #     @
                                              / \
  0 ->                                       C   @
  ^
height
Note: `@` denotes default digest

pub fn get_child_for_proof_and_siblings(
&self,
n: u8
) -> (Option<HashValue>, Vec<HashValue>)

Get the child and its corresponding siblings that are required to generate proof for the n-th child. If it is an existence proof, the returned child must be the n-th child; otherwise, the returned child may be other child. See inline explanation for details. When calling this function with n = 11(b), the range at each level is illustrated as a pair of square brackets:

    4      [f   e   d   c   b   a   9   8   7   6   5   4   3   2   1   0] -> root level
           ---------------------------------------------------------------
    3      [f   e   d   c   b   a   9   8] [7   6   5   4   3   2   1   0] width = 8
                                 chs <--┘                        shs <--┘
    2      [f   e   d   c] [b   a   9   8] [7   6   5   4] [3   2   1   0] width = 4
                 lhs <--┘               └--> chs
    1      [f   e] [d   c] [b   a] [9   8] [7   6] [5   4] [3   2] [1   0] width = 2
                         chs <--┘       └--> shs
    0      [f] [e] [d] [c] [b] [a] [9] [8] [7] [6] [5] [4] [3] [2] [1] [0] width = 1
    ^                chs <--┘   └--> schs
    |   MSB|<---------------------- uint 16 ---------------------------->|LSB
 height    chs: `child_half_start`         shs: `sibling_half_start`