Struct sparse_merkle::SparseMerkleTrie

pub struct SparseMerkleTrie<'a, R: 'a + TrieReader> { /* fields omitted */ }

Methods

impl<'a, R> SparseMerkleTrie<'a, R> where
R: 'a + TrieReader,

pub fn new(reader: &'a R) -> Self

pub fn put_keyed_blob_sets(
&self,
write_sets: Vec<Vec<(HashValue, AccountStateBlob)>>,
root_hash: HashValue
) -> Result<(Vec<HashValue>, TrieUpdateBatch)>

This method will return the new added nodes and blobs of the state in a batch -- the part that have been modified by the write_set. For example, if after transaction T_i the committed state of trie in persistent storage looks like the following structure:

             S_i
            /   \
           .     .
          .       .
         /         \
        o           x
       / \
      A   B
       storage (disk)

where A and B denote the states of two adjacent accounts, and x is a sibling on the path from root to A and B in the trie. Then the next transaction T_{i+1} producing a write_set modified other accounts C and D that lives in the subtree at x, a new partial trie will be constructed in memory and the structure will look like the following:

                S_i      |      S_{i+1}
               /   \     |     /       \
              .     .    |    .         .
             .       .   |   .           .
            /         \  |  /             \
           /           x | /               x'
          o<-------------+-               / \
         / \             |               C   D
        A   B            |
          storage (disk) |    cache (memory)

With this design, we are able to query the global state in persistent storage and produce the proposed trie delta based on a specifit root hash and write_set. For example, if we want to execute another transaction T_{i+1}', we can use the tree S_i in storage and apply the write_set of transaction T_{i+1}. Then if the storage commit the returned batch, the state S_{i+1} is ready to be read from trie api get_with_proof. Anything inside batch is not reachable from public interface before being commited.

pub fn put_keyed_blob_set(
&self,
keyed_blob_set: Vec<(HashValue, AccountStateBlob)>,
root_hash: HashValue
) -> Result<(HashValue, TrieUpdateBatch)>

pub fn get_with_proof(
&self,
key: HashValue,
root_hash: HashValue
) -> Result<(Option<AccountStateBlob>, SparseMerkleProof)>